+++ src/templates/bespoke/bespoke.ts
@@ -12,15 +12,18 @@ import bespokeProgress from './progress'
 import bespokeState from './state'
 import bespokeSync from './sync'
 import bespokeTouch from './touch'
-import { getViewMode, popQuery, setQuery, setViewMode, ViewMode } from './utils'
+import { getViewMode, popQuery, setViewMode, ViewMode } from './utils'
 import bespokeWakeLock from './wake-lock'
 
-const pattern = [ViewMode.Normal, ViewMode.Presenter, ViewMode.Next] as const
-
 const parse = (
-  ...patterns: [[boolean, boolean, boolean], (...args: unknown[]) => void][]
+  ...patterns: [
+    [normalView: boolean, presnterView: boolean, nextView: boolean],
+    (...args: unknown[]) => void
+  ][]
 ) => {
-  const idx = pattern.findIndex((v) => getViewMode() === v)
+  const idx = [ViewMode.Normal, ViewMode.Presenter, ViewMode.Next].findIndex(
+    (v) => getViewMode() === v
+  )
   if (idx < 0) throw new Error('Invalid view')
 
   return patterns.map(([pat, plugin]) => pat[idx] && plugin).filter((p) => p)
@@ -55,11 +58,6 @@ export default function bespokeTemplate(
       [[x, x, _], bespokeWakeLock]
     )
   )
-
-  window.addEventListener('beforeunload', () =>
-    setQuery({ sync: deck.syncKey })
-  )
-  window.addEventListener('unload', () => deck.destroy())
 
   return deck
 }